refactor: 입력 검증 함수 클래스 분리
