+++ src/main/java/lotto/validation/BonusNumberValidator.java
@@ -1,7 +1,5 @@
 package lotto.validation;
 
-import static lotto.validation.CommonLottoNumberValidator.validateCommonLottoInput;
-
 import java.util.List;
 import lotto.domain.Lotto;
 import lotto.view.input.InputErrorMessage;
@@ -10,7 +8,17 @@ import lotto.view.input.InvalidInputException;
 public class BonusNumberValidator {
 
     public static void validateBonusNumber(String input) {
-        validateCommonLottoInput(input);
+        CommonLottoNumberValidator.validateNotNullOrEmpty(input);
+        validateIsNumber(input);
+        CommonLottoNumberValidator.validateLottoRange(Integer.parseInt(input));
+    }
+
+    private static void validateIsNumber(String input) {
+        try {
+            Integer.parseInt(input);
+        } catch (NumberFormatException e) {
+            throw new InvalidInputException(InputErrorMessage.INTEGER_REQUIRED);
+        }
     }
 
     public static void validateDuplicate(Integer input, Lotto winningLotto) {
